requirements-app.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
mlflow
joblib
pydantic
python-multipart
jinja2
python-dotenv